[JSC][GLib] Provide deferred promise creation api https://bugs.webkit.org/show_bug.cgi?id=283708
Reviewed by NOBODY (OOPS!).

This commit implements an API to create promises in JSC GLib.
It is a copy from the Objective-C valueWithNewPromiseInContext.
An executor is required, it will be invoked during promise initialization
and it receives the @resolve and @reject objects than can be called to resolve
or reject the promise. It is called like a js function, so exceptions raised
during the executor invocation will not be propagated to the context, but
handled by the promise causing a rejection.

* Source/JavaScriptCore/API/glib/JSCValue.cpp:
(jsc_value_new_promise):
* Source/JavaScriptCore/API/glib/JSCValue.h.in:
* Tools/TestWebKitAPI/Tests/JavaScriptCore/glib/TestJSC.cpp:
(foooSuccessSource):
(foooSuccessExecutor):
(getAsyncResolvedPromise):
(foooFailureSource):
(foooFailureExecutor):
(getAsyncRejectedPromise):
(getRejectedPromise):
(getExceptionPromise):
(getResolvedPromise):
(testJSCPromises):